// src/api/generated/model/serviceUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * LeadRadar API
 * OpenAPI spec version: 0.1.0
 */

export interface ServiceUpdate {
  name?: string | null;
  description?: string | null;
  value_proposition?: string | null;
  decision_makers?: string[] | null;
  is_active?: boolean | null;
}